conditionally render remove friend button if already friends with that user
