+++ client/src/components/ProfileUserInfo/ProfileUserInfo.jsx
@@ -1,9 +1,9 @@
 import defaultPhoto from '../../assets/images/no-profile-picture.svg'
-import { useMutation } from '@apollo/client';
+import { useMutation, useQuery } from '@apollo/client';
 import { ADD_FRIEND } from '../../utils/mutations';
+import { QUERY_FRIENDS } from '../../utils/queries'
 
 const ProfileUserInfo = ({ data }) => {
-    console.log(data)
     let averageWPM;
     if (data.scoresByUser.length !== 0) {
         let scoresArr = []
@@ -13,16 +13,6 @@ const ProfileUserInfo = ({ data }) => {
         const average = (array) => scoresArr.reduce((a, b) => a + b) / scoresArr.length;
         averageWPM = average(scoresArr);
     }
-    const friendID = data.user._id
-    const [addFriend] = useMutation(ADD_FRIEND);
-
-    const handleFriendSubmit = async (event) => {
-        event.preventDefault();
-        await addFriend({
-            variables: { friendID }
-        });
-        console.log('click')
-    };
 
     return (
         <section>
@@ -46,9 +36,6 @@ const ProfileUserInfo = ({ data }) => {
             <div>
                 {data.user.location && <p>Location: {data.user.location} </p>}
                 {!data.user.location && <p>No location set</p>}
-            </div>
-            <div>
-                <button type="submit" onClick={handleFriendSubmit}>Add friend</button>
             </div>
         </section >
     )
